+++ agent/check_handler.go
@@ -202,7 +202,10 @@ func extractMetrics(event *types.Event) []*types.MetricPoint {
 		transformer, err = transformers.ParseGraphite(event.Check.Output)
 	case types.InfluxDBMetricFormat:
 		transformer, err = transformers.ParseInflux(event.Check.Output)
+	case types.NagiosMetricFormat:
+		transformer, err = transformers.ParseNagios(event)
 	}
+
 	if err != nil {
 		logger.WithError(err).Error("unable to extract metric from check output")
 		return nil